further refactor select builder (less code dup)
